fix(TableGen): change default Jib variable type to uint64_t
Previously, the `translateJibToC` function in EmulatorEmitter defaulted to `uint8_t` when a variable type could not be determined from specific prefixes.

This commit changes the default C++ type to `uint64_t`. This ensures that untyped bitvectors (`%bv`) have sufficient capacity by default, preventing potential truncation issues with variables larger than 8 bits.

+++ llvm/utils/TableGen/EmulatorEmitter.cpp
@@ -426,7 +426,7 @@ static std::string translateJibToC(const JibInstructionBody &Instr,
         if (SanitizedVar == "tmp0")
           continue;
 
-        std::string CppType = "uint8_t"; // Default
+        std::string CppType = "uint64_t"; // Default for untyped %bv
         if (Type.starts_with("%bv16") || Type.starts_with("%i16"))
           CppType = "uint16_t";
         else if (Type.starts_with("%bv32") || Type.starts_with("%i") ||
